# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_api.CreateUserRequest:
    properties:
      password:
        example: secure_password
        type: string
      pin:
        example: "1234"
        type: string
      roles:
        example:
        - '[''user'''
        - '''admin'']'
        items:
          type: string
        type: array
      username:
        example: john.doe
        type: string
    required:
    - password
    - username
    type: object
...
  internal_api.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
  internal_api.LoginRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      password:
        example: secure_password
        type: string
    required:
    - email
    - password
    type: object
  internal_api.LoginResponse:
    properties:
      refresh_token:
        example: ""
        type: string
      token:
        example: eyJhbGciOiJS...
        type: string
      user:
        properties:
          username:
            example: john.doe@example.com
            type: string
        type: object
    type: object
  internal_api.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  internal_api.TransferRequest:
    properties:
      amount:
        example: 1000
        type: integer
      is_anonymous:
        example: false
        type: boolean
      pin:
        example: "1234"
        type: string
      receiver_wallet_id:
        example: 2
        type: integer
      sender_wallet_id:
        example: 1
        type: integer
    required:
    - amount
    - receiver_wallet_id
    - sender_wallet_id
    type: object
...
  internal_api.UpdateUserRequest:
    properties:
      display_name:
        example: John Doe
        type: string
      pin_required_for_transfer:
        example: true
        type: boolean
      profile_photo_url:
        example: https://example.com/photo.jpg
        type: string
    type: object
...
  verve_internal_models.Transfer:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      is_anonymous:
        type: boolean
      receiver_wallet_id:
        type: integer
      sender_wallet_id:
        type: integer
      status:
        $ref: '#/definitions/verve_internal_models.TransferStatus'
      updated_at:
        type: string
    type: object
  verve_internal_models.TransferStatus:
    enum:
    - pending
    - completed
    - failed
    type: string
    x-enum-varnames:
    - TransferStatusPending
    - TransferStatusCompleted
    - TransferStatusFailed
  verve_internal_models.User:
    description: User information including display name and profile settings
    properties:
      created_at:
        type: string
      display_name:
        example: John Doe
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 1
        type: integer
      pin_required_for_transfer:
        example: true
        type: boolean
      profile_photo_url:
        example: https://example.com/photo.jpg
        type: string
      provider:
        description: OAuth provider (google, okta, or local)
        example: google
        type: string
      provider_user_id:
        description: ID from the OAuth provider
        type: string
      roles:
        example:
        - '[''user'''
        - '''admin'']'
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        example: john.doe
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Verve API
  version: "1.0"
paths:
  /auth/link/oauth:
    post:
      consumes:
      - application/json
      description: Link an OAuth account to the current user's account
      parameters:
      - description: OAuth provider (google/okta)
        in: query
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Link OAuth account
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/unlink/oauth:
    post:
      description: Remove OAuth provider from user account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.SuccessResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Unlink OAuth account
      tags:
      - auth
...
  /transfer:
    post:
      consumes:
      - application/json
      description: Create a new transfer between wallets
      parameters:
      - description: Transfer details
        in: body
        name: transfer
        required: true
        schema:
          $ref: '#/definitions/internal_api.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.Transfer'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Insufficient funds or invalid PIN
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Initiate a transfer
      tags:
      - transfers
  /transfer/{id}:
    get:
      description: Get the current status of a transfer
      parameters:
      - description: Transfer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.Transfer'
        "400":
          description: Invalid transfer ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: Transfer not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get transfer status
      tags:
      - transfers
  /user/{id}:
    get:
      description: Get details of a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user details
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile information
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User details to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_api.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Can only update own profile
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update user profile
      tags:
      - users
...
  /user/connected:
    get:
      description: Get a list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/verve_internal_models.User'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "404":
          description: No users found
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List all users
      tags:
      - users
  /user/register:
    post:
      consumes:
      - application/json
      description: Create a new user with specified roles (admin only)
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/internal_api.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/verve_internal_models.User'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/internal_api.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create new user
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"